# docs/swagger.yaml
definitions:
  handler.CreateAccountRequest:
    properties:
      document_number:
        type: string
    type: object
  handler.CreateAccountResponse:
    properties:
      message:
        type: string
    type: object
  handler.CreateTransactionRequest:
    properties:
      account_id:
        type: integer
      amount:
        type: number
      operation_type_id:
        type: integer
    type: object
  handler.CreateTransactionResponse:
    properties:
      message:
        type: string
    type: object
  handler.ErrorResponse:
    properties:
      errorCode:
        type: string
      message:
        type: string
    type: object
  handler.ShowAccountResponse:
    properties:
      account_id:
        type: integer
      document_number:
        type: string
    type: object
info:
  contact: {}
paths:
  /accounts:
    post:
      consumes:
      - application/json
      description: Create a new account
      parameters:
      - description: Request body (DocumentNumber must be an 11 character long string)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.CreateAccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create an account
      tags:
      - Accounts
  /accounts/{accountId}:
    get:
      consumes:
      - application/json
      description: Find an account by its ID
      parameters:
      - description: Account ID
        in: path
        name: accountId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ShowAccountResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Show an account
      tags:
      - Accounts
  /transactions:
    post:
      consumes:
      - application/json
      description: Create a new transaction
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateTransactionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.CreateTransactionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create a transaction
      tags:
      - Transactions
swagger: "2.0"